requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
babel==2.17.0
black==25.1.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
courlan==1.3.2
cryptography==45.0.5
cymem==2.0.11
dateparser==1.2.2
dgl==2.2.0
distlib==0.3.9
fastapi==0.115.14
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.5
htmldate==1.9.3
httptools==0.6.4
huggingface-hub==0.33.2
identify==2.6.12
idna==3.10
iniconfig==2.1.0
interchange==2021.0.4
isort==6.0.1
Jinja2==3.1.6
jusText==3.0.2
langcodes==3.5.0
language_data==1.3.0
lxml==5.4.0
lxml_html_clean==0.4.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.6.3
murmurhash==1.0.13
mypy_extensions==1.1.0
neo4j==5.28.1
networkx==3.5
nodeenv==1.9.1
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pansi==2024.11.0
pathspec==0.12.1
pdfminer.six==20221105
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
preshed==3.0.10
propcache==0.3.2
psutil==7.0.0
py2neo==2021.2.4
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
ruff==0.12.1
safetensors==0.5.3
scipy==1.16.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.2
sympy==1.14.0
thinc==8.3.6
tld==0.13.1
tokenizers==0.21.2
torch==2.7.1
torch-geometric==2.6.1
torchdata==0.11.0
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.53.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.31.2
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
scikit-learn==1.4.2
sentence-transformers==2.7.0
